+++ packages/core/src/odata/common/request/index.ts
@@ -4,4 +4,3 @@ export * from './function-import-parameter';
 export * from './odata-request';
 export * from './odata-request-config';
 export * from './odata-request-traits';
-export * from './uri-converter';