+++ packages/core/src/odata/common/uri-conversion/index.ts
@@ -5,3 +5,4 @@ export * from './get-keys';
 export * from './get-orderby';
 export * from './get-resource-path';
 export * from './odata-uri';
+export { UriConverter } from './uri-value-converter';