+++ packages/core/src/odata/common/uri-conversion/odata-uri.ts
@@ -9,7 +9,8 @@ import { Orderable } from '../order';
 import { EdmTypeShared } from '../edm-types';
 
 /**
- * @experimental This is experimental and is subject to change. Use with caution.
+ * Union of necessary methods for the OData URI conversion.
+ * In v2/uri-conversion/odata-uri.ts and v4/uri-conversion/odata-uri.ts the instance for v2 and v4 are created.
  */
 export interface ODataUri {
   getExpand<EntityT extends EntityBase>(
@@ -35,7 +36,10 @@ export interface ODataUri {
   getSelect<EntityT extends EntityBase>(
     selects: Selectable<EntityT>[]
   ): Partial<{ select: string }>;
-  convertToUriFormat(value: any, edmType: EdmTypeShared<'any'>): string;
+  convertToUriFormat(
+    value: any,
+    edmType: EdmTypeShared<'v2'> | EdmTypeShared<'v4'>
+  ): string;
 }
 
 export function prependDollar(param: string): string {