+++ packages/core/src/odata/v2/selectable/custom-field.ts
@@ -2,7 +2,7 @@
 
 import type { Entity } from '../entity';
 import { CustomFieldBase } from '../../common/selectable/custom-field';
-import { DateField } from '../../common';
+import { DateField, TimeField } from '../../common';
 
 export class CustomField<EntityT extends Entity> extends CustomFieldBase<
   EntityT
@@ -12,6 +12,13 @@ export class CustomField<EntityT extends Entity> extends CustomFieldBase<
       this._fieldName,
       this._entityConstructor,
       'Edm.DateTime'
+    );
+  }
+  edmTime(): TimeField<EntityT> {
+    return new TimeField<EntityT>(
+      this._fieldName,
+      this._entityConstructor,
+      'Edm.Time'
     );
   }
 }