+++ packages/core/src/odata/v4/filter/date-filter-function.ts
@@ -1,9 +1,9 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
 
 import moment from 'moment';
-import { EntityBase } from '../common';
-import { FilterFunctionParameterType } from '../common/filter';
-import { OrderableFilterFunction } from '../common/filter/orderable-filter-function';
+import { EntityBase } from '../../common';
+import { FilterFunctionParameterType } from '../../common/filter';
+import { OrderableFilterFunction } from '../../common/filter/orderable-filter-function';
 
 /**
  * Representation of a filter function, that returns a value of type date. This supports DateTimeOffset values.