+++ packages/core/src/odata/v4/request-builder/action-import-request-builder.ts
@@ -7,9 +7,9 @@ import {
 import { ActionFunctionImportRequestBuilderBase } from '../../common/request-builder/action-function-import-request-builder-base';
 
 /**
- * Create OData request to execute a action import.
- * @typeparam ParametersT - Type of the function import parameters
- * @typeparam ReturnT - Type of the function import return value
+ * Create an OData request to execute an action import.
+ * @typeparam ParametersT - Type of the action import parameters
+ * @typeparam ReturnT - Type of the action import return value
  */
 export class ActionImportRequestBuilder<
   ParametersT,
@@ -18,7 +18,7 @@ export class ActionImportRequestBuilder<
   /**
    * Creates an instance of ActionImportRequestBuilder.
    * @param defaultServicePath - Default path for the service the action belongs to
-   * @param actionImportName - The name of the function import.
+   * @param actionImportName - The name of the action import.
    * @param responseTransformer - Transformation function for the response
    * @param parameters - Parameters to be set in the action
    */