+++ packages/core/src/odata/v4/request-builder/get-by-key-request-builder.ts
@@ -21,7 +21,9 @@ import { oDataUri } from '../uri-conversion';
 import { HttpReponse } from '../../../http-client';
 import { getSingleResult } from './response-data-accessor';
 /**
- * Create OData request to get a single entity based on its key properties. A `GetByKeyRequestBuilder` allows to restrict the response to a selection of fields,
+ * Create an OData request to get a single entity based on its key properties.
+ * The properties available in the response can be restricted by creating a [[GetByKeyRequestBuilder.select selection]], where no selection is equal to selecting all fields of the entity.
+ * Navigational properties need to expanded explicitly by [[GetAllRequestBuilder.expand]].
  * where no selection is equal to selecting all fields.
  *
  * @typeparam EntityT - Type of the entity to be requested