+++ packages/core/src/odata/v4/request-builder/response-data-accessor.ts
@@ -1,17 +1,32 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
 
-import { createLogger } from '@sap-cloud-sdk/util';
+import { createLogger, MapType } from '@sap-cloud-sdk/util';
 
 const logger = createLogger({
   package: 'core',
   messageContext: 'response-data-accessor'
 });
 
+/**
+ * Methods to extract the data from OData v4 responses.
+ */
+
+/**
+ * Extract the collection data from the response.
+ * If the data does not contain a collection an empty array is returned.
+ * @param data - Response of the OData v4 service
+ * @returns any[] - Collection extracted from the response
+ */
 export function getCollectionResult(data): any[] {
   validateCollectionResult(data);
   return isCollectionResult(data) ? data.value : [];
 }
 
+/**
+ * Checks if the data contains a collection result.
+ * @param data - Response of the OData v4 service
+ * @returns boolean - true if the data is a collection result
+ */
 export function isCollectionResult(data): boolean {
   return Array.isArray(data.value);
 }
@@ -24,7 +39,13 @@ function validateCollectionResult(data): void {
   }
 }
 
-export function getSingleResult(data): Record<string, any> {
+/**
+ * Extract the single entry data from the response.
+ * If the data does not contain a single object an empty object is returned.
+ * @param data - Response of the OData v4 service
+ * @returns MapType<any> - single object extracted from the response
+ */
+export function getSingleResult(data): MapType<any> {
   validateSingleResult(data);
   return isSingleResult(data) ? data : {};
 }