+++ packages/core/src/odata/v4/uri-conversion/get-expand.ts
@@ -2,15 +2,29 @@
 
 import { Entity } from '../entity';
 import { Expandable } from '../../common/expandable';
-import { Constructable, AllFields, Link, and } from '../../common';
+import {
+  Constructable,
+  AllFields,
+  Link,
+  and,
+  createGetFilter
+} from '../../common';
 import { OneToManyLink } from '../../common/selectable/one-to-many-link';
-import { getFilter } from './get-filter';
 import { getSelect } from './get-select';
+import { uriConverter } from './uri-value-converter';
 
 function prependDollar(param: string): string {
   return `$${param}`;
 }
 
+/**
+ * Get an object containing the given expand as a query parameter, or an empty object if none was given.
+ *
+ * @typeparam EntityT - Type of the entity to expand on
+ * @param expands - The expands to transform to a query parameter
+ * @param entityConstructor - Constructor type of the entity to expand on
+ * @returns An object containing the query parameter or an empty object
+ */
 export function getExpand<EntityT extends Entity>(
   expands: Expandable<EntityT>[] = [],
   entityConstructor: Constructable<EntityT>
@@ -44,7 +58,10 @@ function getExpandAsString<EntityT extends Entity>(
     if (expand instanceof OneToManyLink) {
       params = {
         ...params,
-        ...getFilter(and(...expand._filters?.filters), entityConstructor),
+        ...createGetFilter(uriConverter).getFilter(
+          and(...expand._filters?.filters),
+          entityConstructor
+        ),
         ...(expand._skip && { skip: expand._skip }),
         ...(expand._top && { top: expand._top })
       };