+++ packages/core/test/request-builder/delete-request-builder-v4.spec.ts
@@ -1,7 +1,7 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
 import nock from 'nock';
 import { v4 as uuid } from 'uuid';
-import { convertToUriFormat, DeleteRequestBuilder } from '../../src/odata/v4';
+import { DeleteRequestBuilder, uriConverter } from '../../src/odata/v4';
 import { muteLoggers } from '../test-util/mute-logger';
 import {
   defaultDestination,
@@ -9,6 +9,8 @@ import {
 } from '../test-util/request-mocker';
 import { testEntityResourcePath } from '../test-util/test-data';
 import { TestEntity } from '../test-util/test-services/v4/test-service';
+
+const { convertToUriFormat } = uriConverter;
 
 describe('DeleteRequestBuilder', () => {
   const keyPropGuid = uuid();