+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Added experimental support for OData V4 enum type. Use the `@sap/cloud-sdk-generator` to generate the type safe client.
 - Include fallback `Edm.Any` for previously unsupported edm types like Edm.Geography.
 Type safety and related features like URI conversion are not supported for the fallback type.  
+- OData v4 is moved from the experimental state to GA.
 
 ## Improvements